brain/pyproject.toml
[project]
name = "dina-brain"
version = "0.4.0"
description = "Dina Brain — LLM reasoning sidecar for dina-core"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "httpx>=0.27.0",
    "structlog>=24.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
llm = ["google-generativeai>=0.8.0", "anthropic>=0.40.0"]
pii = ["spacy>=3.7.0"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24.0", "pytest-cov"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
